Change lighting value from float to half for performance

+++ ShaderLibrary/Surface.hlsl
@@ -6,10 +6,10 @@ struct Surface
     float3 position;
     float3 viewDir;
     float3 normal;
-    float3 albedo;
-    float metallic;
-    float roughness;
-    float ao;
+    half3 albedo;
+    half metallic;
+    half roughness;
+    half ao;
 };
 
 #endif
